requirements.txt
kafka-python
mysql-connector-python
pymysql

